requirements.txt
attrs==23.2.0
boto3==1.34.155
botocore==1.34.155
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
h11==0.14.0
idna==3.7
jmespath==1.0.1
numpy==2.0.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
